pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "coderAI"
version = "0.2.0"
description = "Intelligent coding agent CLI with a React-powered terminal UI (Ink) and MCP tools"
readme = "README.md"
requires-python = ">=3.9"
authors = [
    {name = "CoderAI", email = "hello@coderai.dev"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "rich>=13.7.0",
    "click>=8.1.7",
    "openai>=1.10.0",
    "requests>=2.31.0",
    "pydantic>=2.5.0",
    "aiohttp>=3.9.0",
    "tiktoken>=0.5.2",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "groq>=0.4.2",
    "certifi>=2024.2.2",
    "chromadb>=0.4.22",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "types-PyYAML>=6.0.0",
]

[project.urls]
Homepage = "https://github.com/adityaanilraut/CoderAI"
Releases = "https://github.com/adityaanilraut/CoderAI/releases"
Issues = "https://github.com/adityaanilraut/CoderAI/issues"
Source = "https://github.com/adityaanilraut/CoderAI"

[project.scripts]
coderAI = "coderAI.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["coderAI*"]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311", "py312"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint.per-file-ignores]
# load_dotenv() must run before package imports read env vars
"coderAI/cli.py" = ["E402"]
"coderAI/ipc/entry.py" = ["E402"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
exclude = ["tests/"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
asyncio_default_fixture_loop_scope = "function"

